gcf/ingest-audio/requirements.txt
functions-framework==3.*
google-cloud-storage==2.18.2
grpcio==1.66.1
grpcio-tools==1.66.1
requests==2.32.3